style(global.css): Adding custom fonts, assign to html elements.
