+++ src/components/Header/index.js
@@ -1,2 +1,2 @@
-export { default } from './Header';
+export { default } from './Header.component';
