+++ src/components/Layout/index.js
@@ -1,2 +1,2 @@
-export { default } from './Layout';
+export { default } from './Layout.component';
