+++ src/global.css
@@ -1,3 +1,13 @@
+@font-face {
+  font-family: Ashbury;
+  src: local("Ashbury"), url("./assets/fonts/ashbury-light.otf") format("opentype");
+}
+
+@font-face {
+  font-family: Graphik;
+  src: local("Graphik"), url("./assets/fonts/graphik-regular.otf") format("opentype");
+}
+
 html {
   font-size: 1.125rem;
   line-height: 1.6;
@@ -26,6 +36,7 @@ html {
 body {
   margin: 0;
   font-family:
+    sans-serif,
     -apple-system,
     BlinkMacSystemFont,
     "Segoe UI",
@@ -35,8 +46,7 @@ body {
     Cantarell,
     "Fira Sans",
     "Droid Sans",
-    "Helvetica Neue",
-    sans-serif;
+    "Helvetica Neue";
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
@@ -58,4 +68,41 @@ section {
 
 figure {
   margin: 0;
+}
+
+h1,
+h2 {
+  font-family: Ashbury, sans-serif;
+  margin-top: -0.095em;
+  margin-bottom: 0.095em;
+  padding-bottom: 1.75rem;
+  max-width: 40em;
+}
+
+h4,
+h5 {
+  font-family: Graphik, sans-serif;
+}
+
+h1 {
+  font-weight: 500;
+}
+
+h2,
+h3 {
+  font-weight: 400;
+}
+
+span {
+  font-family: Graphik, sans-serif;
+  font-weight: 400;
+  font-size: 0.875rem;
+  line-height: 1.5rem;
+}
+
+p {
+  font-family: Ashbury, sans-serif;
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.5rem;
 }
